+++ src/lib.rs
@@ -1,0 +1,11 @@
+#![recursion_limit = "1024"]
+extern crate getopts;
+extern crate quote;
+extern crate syn;
+#[macro_use]
+extern crate log;
+extern crate env_logger;
+extern crate failure;
+
+mod util;
+pub use util::create_directory_structure;
